Move accumulators to own file
